feat: refactor Account page to use SectionAccount and SectionProfile components for better organization and maintainability
